Fix/compat rax element (#5847)
* fix: dynamic import works in app.tsx

* fix: test case

* chore: update depenedency version of @ice/webpack-config

* refactor: rax-compat jsx-runtime

* chore: add CHANGELOG and update version of rax-compat

* chore: pnpm lock up-to-date

* fix: puppteer update
