+++ packages/rax-compat/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.7
+
+- [refactor] jsx-runtime using react backend implementation.
+
 ## 0.1.6
 
 - [feat] provide jsx-runtime for JSX transform automatic.
