+++ packages/rax-compat/package.json
@@ -1,6 +1,6 @@
 {
   "name": "rax-compat",
-  "version": "0.1.6",
+  "version": "0.1.7",
   "description": "Rax compatible mode, running rax project on the react runtime.",
   "files": [
     "esm",
