+++ packages/rax-compat/src/runtime/jsx-runtime.ts
@@ -1,1 +1,7 @@
-export { jsx, jsxs, Fragment } from './impl.js';
+// @ts-ignore
+import { jsxs as _jsxs, jsx as _jsx } from 'react/jsx-runtime';
+import { createJSXElementFactory } from '../compat/element';
+
+export { Fragment } from 'react';
+export const jsx = createJSXElementFactory(_jsx);
+export const jsxs = createJSXElementFactory(_jsxs);
