+++ packages/rax-compat/src/style.ts
@@ -1,0 +1,39 @@
+import { cached, convertUnit } from 'style-unit';
+import { isNumber, isObject } from './type.js';
+
+// https://github.com/alibaba/rax/blob/master/packages/driver-dom/src/index.js
+// opacity -> opa
+// fontWeight -> ntw
+// lineHeight|lineClamp -> ne[ch]
+// flex|flexGrow|flexPositive|flexShrink|flexNegative|boxFlex|boxFlexGroup|zIndex -> ex(?:s|g|n|p|$)
+// order -> ^ord
+// zoom -> zoo
+// gridArea|gridRow|gridRowEnd|gridRowSpan|gridRowStart|gridColumn|gridColumnEnd|gridColumnSpan|gridColumnStart -> grid
+// columnCount -> mnc
+// tabSize -> bs
+// orphans -> orp
+// windows -> ows
+// animationIterationCount -> onit
+// borderImageOutset|borderImageSlice|borderImageWidth -> erim
+const NON_DIMENSIONAL_REG = /opa|ntw|ne[ch]|ex(?:s|g|n|p|$)|^ord|zoo|grid|orp|ows|mnc|^columns$|bs|erim|onit/i;
+const isDimensionalProp = cached((prop: string) => !NON_DIMENSIONAL_REG.test(prop));
+
+// Convert unit as driver-dom does.
+// https://github.com/alibaba/rax/blob/master/packages/driver-dom/src/index.js#L346
+export function compatStyle<S = object>(style?: S): S | void {
+  if (isObject(style)) {
+    // Do not modify the original style object, copy results to another plain object.
+    const result = Object.create(Object.prototype);
+    for (let key in style) {
+      const value = style[key];
+      if (isNumber(value) && isDimensionalProp(key)) {
+        // Transform rpx to vw.
+        result[key] = convertUnit(`${value}rpx`);
+      } else {
+        result[key] = convertUnit(value);
+      }
+    }
+    return result;
+  }
+  return style;
+}
